+++ eskema/fsspec/monkey.py
@@ -1,7 +1,0 @@
-import fsspec.spec
-
-from eskema.fsspec.spec import readlines
-
-
-def activate():
-    fsspec.spec.AbstractBufferedFile.readlines = readlines