+++ tests/test_csv_infer.py
@@ -1,5 +1,3 @@
-import io
-
 import pytest
 from click.testing import CliRunner
 
@@ -7,20 +5,6 @@ from eskema.cli import cli
 from eskema.core import SchemaGenerator
 from eskema.model import Resource, SqlResult, SqlTarget
 from tests.util import BACKENDS, get_basic_sql_reference, getcmd
-
-
-@pytest.fixture
-def basic_stream_csv():
-    """
-    A stream of input data. Here, in CSV format.
-    """
-    return io.StringIO(
-        """
-id,name,date,fruits,price
-1,"foo","2014-10-31T09:22:56","apple,banana",0.42
-2,"bar",,"pear",0.84
-    """.strip()
-    )
 
 
 @pytest.mark.parametrize("backend", BACKENDS)