+++ tests/test_json_document_autopk.py
@@ -3,7 +3,7 @@ from click.testing import CliRunner
 from eskema.autopk import infer_pk
 from eskema.cli import cli
 from eskema.type import ContentType
-from eskema.util import sql_canonicalize
+from eskema.util.sql import sql_canonicalize
 from tests.util import get_basic_sql_reference, random_table_name