+++ tests/test_lineprotocol.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 from eskema.cli import cli
 from eskema.core import SchemaGenerator
 from eskema.model import Resource, SqlResult, SqlTarget
-from eskema.util import unwrap
+from eskema.util.data import unwrap
 from tests.util import getcmd
 
 reference = unwrap(