+++ CHANGES.rst
@@ -40,6 +40,7 @@ in progress
 - Refactoring: Rework ``eskema.type``
 - Refactoring: Add ``eskema.io.open`` as a wrapper around ``fsspec.open``
 - Refactoring: Add ``eskema.io.to_dataframe`` from ``eskema.autopk``
+- Refactoring: Add ``eskema.util`` folder instead of single ``util.py``
 
 
 2023-01-xx 0.0.0